+++ interceptor/handler/static.go
@@ -3,6 +3,7 @@ package handler
 import (
 	"net/http"
 
+	kedahttp "github.com/kedacore/http-add-on/pkg/http"
 	"github.com/kedacore/http-add-on/pkg/k8s"
 	"github.com/kedacore/http-add-on/pkg/routing"
 	"github.com/kedacore/http-add-on/pkg/util"
@@ -18,6 +19,10 @@ func NewStatic(statusCode int, err error) http.Handler {
 		stream := util.UpstreamURLFromContext(ctx)
 
 		statusText := http.StatusText(statusCode)
+		if statusText == "" && statusCode == kedahttp.StatusClientClosedRequest {
+			// 499 isn't a registered IANA status code, so http.StatusText doesn't know it.
+			statusText = "Client Closed Request"
+		}
 		routingKey := routing.NewKeyFromRequest(r)
 		namespacedName := k8s.NamespacedNameFromObject(ir)
 		logger.Error(err, statusText, "routingKey", routingKey, "namespacedName", namespacedName, "stream", stream)